Factorize and remove code
