+++ src/Mining/CoinbaseField/index.js
@@ -14,17 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-import React from 'react';
-import PropTypes from 'prop-types';
+import CoinbaseField from './CoinbaseField';
 
-import Header from 'semantic-ui-react/dist/commonjs/elements/Header';
-import Segment from 'semantic-ui-react/dist/commonjs/elements/Segment';
-
-const Section = ({ children, title, ...rest }) => (
-  <Segment padded {...rest}>
-    <Header as="h3">{title}</Header>
-    {children}
-  </Segment>
-);
-
-export default Section;
+export default CoinbaseField;
