+++ src/Mining/Mining.js
@@ -19,25 +19,20 @@ import PropTypes from 'prop-types';
 import { inject, observer } from 'mobx-react';
 import Form from 'semantic-ui-react/dist/commonjs/collections/Form';
 import Statistic from 'semantic-ui-react/dist/commonjs/views/Statistic';
-import IdentityIcon from '@parity/ui/lib/IdentityIcon';
-import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 import formatNumber from 'format-number';
 
 import Field from '../components/Field';
+import LowerCaseStatistic from '../components/LowerCaseStatistic';
 import Section from '../components/Section';
+import CoinbaseField from './CoinbaseField/CoinbaseField';
 import decodeExtraData from './utils/decodeExtraData';
 import numberFromString from './utils/numberFromString';
-import styles from './Mining.css';
 
 const toNiceNumber = formatNumber();
 
 class Mining extends Component {
-  static propTypes = {
-    intl: intlShape
-  };
-
-  handleCoinbaseChange = (_, { value }) =>
-    this.props.coinbaseStore.setAuthor(value);
+  static propTypes = {};
 
   handleGasFloorTargetChange = value =>
     this.props.gasFloorTargetStore.saveGasFloorTarget(numberFromString(value));
@@ -50,13 +45,10 @@ class Mining extends Component {
 
   render() {
     const {
-      accountsStore: { accounts },
-      coinbaseStore: { coinbase },
       extraDataStore: { extraData },
       gasFloorTargetStore: { gasFloorTarget },
       hashrateStore: { hashrate },
       latestBlockStore: { latestBlock },
-      intl: { formatMessage },
       minGasPriceStore
     } = this.props;
 
@@ -70,26 +62,7 @@ class Mining extends Component {
         }
       >
         <Form>
-          <Form.Select
-            icon={
-              coinbase && (
-                <IdentityIcon
-                  address={coinbase}
-                  alt={coinbase}
-                  className={styles.coinbaseAvatar}
-                />
-              )
-            }
-            label={formatMessage(messages.coinbaseLabel)}
-            onChange={this.handleCoinbaseChange}
-            options={accounts.map(account => ({
-              image: <IdentityIcon address={account} alt={account} />,
-              key: account,
-              value: account,
-              text: account
-            }))}
-            value={coinbase}
-          />
+          <CoinbaseField />
           <Field
             label={
               <FormattedMessage
@@ -121,16 +94,12 @@ class Mining extends Component {
             value={toNiceNumber(gasFloorTarget)}
           />
         </Form>
-        <Statistic.Group
-          size="tiny"
-          widths={2}
-          className={styles.smallStatistic}
-        >
+        <LowerCaseStatistic size="tiny" widths={2}>
           <Statistic>
             <Statistic.Value>{`#${toNiceNumber(
               latestBlock.number || ' -'
             )}`}</Statistic.Value>
-            <Statistic.Label className={styles.lowerCaseLabel}>
+            <Statistic.Label>
               <FormattedMessage
                 id="dapp.status.mining.bestBlockLabel"
                 defaultMessage="Best Block, at {time}"
@@ -143,36 +112,26 @@ class Mining extends Component {
             </Statistic.Label>
           </Statistic>
           <Statistic>
-            <Statistic.Value className={styles.lowerCaseLabel}>{`${hashrate ||
-              0} H/s`}</Statistic.Value>
-            <Statistic.Label className={styles.lowerCaseLabel}>
+            <Statistic.Value>{`${hashrate || 0} H/s`}</Statistic.Value>
+            <Statistic.Label>
               <FormattedMessage
                 id="dapp.status.mining.hashrateLabel"
                 defaultMessage="Hash Rate"
               />
             </Statistic.Label>
           </Statistic>
-        </Statistic.Group>
+        </LowerCaseStatistic>
       </Section>
     );
   }
 }
 
-export default injectIntl(
-  inject(
-    'accountsStore',
-    'coinbaseStore',
-    'extraDataStore',
-    'gasFloorTargetStore',
-    'hashrateStore',
-    'latestBlockStore',
-    'minGasPriceStore'
-  )(observer(Mining))
-);
-
-const messages = {
-  coinbaseLabel: {
-    id: 'dapp.status.mining.coinbaseLabel',
-    defaultMessage: 'Mining Author'
-  }
-};
+export default inject(
+  'accountsStore',
+  'coinbaseStore',
+  'extraDataStore',
+  'gasFloorTargetStore',
+  'hashrateStore',
+  'latestBlockStore',
+  'minGasPriceStore'
+)(observer(Mining));
