+++ src/components/LowerCaseStatistic/LowerCaseStatistic.css
@@ -15,15 +15,10 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-:local(.coinbaseAvatar) {
-  width: 1em;
-  height: 1em;
-}
-
-:local(.smallStatistic) {
+:local(.lowerCaseStatistic) {
   margin-top: 2em !important;
 }
 
-:local(.lowerCaseLabel) {
+:local(.lowerCaseStatistic) .label {
   text-transform: none !important;
 }
